fix(mac): store mutliple formats in one NSPasteboardItem
On MacOS, the `NSPasteboardItem` can store multiple formats of the same data.
The previous implementation created a seperate NSParseboardItem for each format - which contextually
treats them as different items not different formats of the same item.

The consequence of this is that when setting Text and HTML and then
pasting, the text is duplicated because two different items are pasted
instead of just one format of the same item.